UCF: run scales D_led

+++ ucf_sports/dim_14/classification/1-NN/cross-validation-NN-def.hpp
@@ -6,7 +6,7 @@ class cv_classify_NN //KTH Cross validation
     inline cv_classify_NN(const std::string in_path,
 			  const std::string in_path_dataset,
 			  const std::string in_actionNames,  
-			  const int in_scale_factor, 
+			  const float in_scale_factor, 
 			  const int in_shift,
 			  const int in_dim 
 			 );
